fix validator reward cache circular dependency
